Refactor various components to ensure proper type casting and improve code clarity
- Updated rowCount and count methods in ChatMessageListModel and AgentActivityLog to use static_cast for size conversions.
- Introduced a utility function toInt for consistent conversion from qsizetype to int across multiple files.
- Enhanced the applyVoiceConfigurationPaths method in DesktopShellViewModel with comments for clarity.
- Added a None value to the RetrievalSourcePriority enum for better handling of retrieval priorities.
- Refactored InMemoryConversationStore and InMemoryMemoryCandidateStore to ensure proper size handling.
- Adjusted InMemorySettingsStore, InMemoryStore, and InMemoryToolRegistry to avoid unnecessary move semantics.
- Improved test cases in InMemoryToolRegistryTest for better error handling and clarity.

+++ apps/sentinel-desktop/src/ChatMessageListModel.cpp
@@ -5,7 +5,7 @@ namespace sentinel::desktop {
 ChatMessageListModel::ChatMessageListModel(QObject* parent) : QAbstractListModel(parent) {}
 
 int ChatMessageListModel::rowCount(const QModelIndex& parent) const {
-    return parent.isValid() ? 0 : messages_.size();
+    return parent.isValid() ? 0 : static_cast<int>(messages_.size());
 }
 
 QVariant ChatMessageListModel::data(const QModelIndex& index, int role) const {
